feat: more groundwork
